Added skips for conduit tests when library doesn't exist

+++ tests/data_formatting/conduit/test_conduit.py
@@ -3,15 +3,14 @@ import sys
 
 import numpy as np
 
-
+skip_conduit_tests = False
 
 try:
     import conduit
     from spellbook.data_formatting.conduit.python import conduit_bundler as cb
 except ModuleNotFoundError:
-    print("Conduit not available! These tests will fail!")
-    print("Exiting nicely!")
-    sys.exit()
+    print("Conduit not available! These tests will be skipped!")
+    skip_conduit_tests = True
 
 def make_dummy_node():
     x = conduit.Node()
@@ -23,10 +22,14 @@ def make_dummy_node():
 
 
 def test_make_conduit_node():
+    if skip_conduit_tests:
+        return
     x = make_dummy_node()
 
 
 def test_save_node():
+    if skip_conduit_tests:
+        return
     x = make_dummy_node()
     save_node_many(x)
     delete_data()
@@ -47,6 +50,8 @@ def delete_conduit_file(filename):
 
 
 def test_load_node():
+    if skip_conduit_tests:
+        return
     x = make_dummy_node()
     save_node_many(x)
     _ = load_node_many(base="_dummy")
@@ -54,6 +59,8 @@ def test_load_node():
 
 
 def test_load_handle():
+    if skip_conduit_tests:
+        return
     base = "_dummy"
     exts = (".h5", ".hdf5", ".json", ".yaml", ".cbin")
     x = make_dummy_node()
